requirements.txt
torch
transformers
Pillow
runpod
sentencepiece
torchvision